feat: move extensions tool
