+++ codex-rs/core/src/tools/handlers/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod agent_jobs_spec;
 pub(crate) mod apply_patch;
 pub(crate) mod apply_patch_spec;
 mod dynamic;
+pub(crate) mod extension_tools;
 mod goal;
 pub(crate) mod goal_spec;
 mod mcp;
